refactor: LottoListDisplay 분리
