+++ src/components/LottoListDisplay/elements.js
@@ -1,0 +1,3 @@
+export const lottoStatus = document.querySelector("#lotto-status");
+export const lottoTicketList = document.querySelector("#lotto-ticket-list");
+export const lottoTicketTemplate = document.querySelector("#lotto-ticket");
